+++ 05-JWT-Basics/starter/controllers/main.js
@@ -28,24 +28,13 @@ const login = async (req, res) => {
 };
 
 const dashboard = async (req, res) => {
-    const authHeader = req.headers.authorization;
+    console.log(req.user);
+    const luckyNumber = Math.floor(Math.random() * 100);
 
-    if (!authHeader || !authHeader.startsWith("Bearer ")) {
-      throw new CustomAPIError("No token provided", 401);
-    }
-
-    const token = authHeader.split(" ")[1];
-
-    try {
-      const decoded = jwt.verify(token, process.env.JWT_SECRET);
-      const luckyNumber = Math.floor(Math.random() * 100);
-      res.status(200).json({
-        message: `Hello, ${decoded.username}`,
-        secret: `Here is your authorized data, and your lucky number is ${luckyNumber}`,
-      });
-    } catch (error) {
-      throw new CustomAPIError("Not authorized to access this route", 401);
-    }
+    res.status(200).json({
+      msg: `Hello, ${req.user.username}`,
+      secret: `Here is your authorized data, and your lucky number is ${luckyNumber}`,
+    });
 };
 
 module.exports = { login, dashboard };
