+++ 05-JWT-Basics/starter/routes/main.js
@@ -3,7 +3,9 @@ const router = express.Router();
 
 const { login, dashboard } = require("../controllers/main");
 
-router.route("/dashboard").get(dashboard);
+const authMiddleware = require("../middleware/auth");
+
+router.route("/dashboard").get(authMiddleware, dashboard);
 router.route("/login").post(login);
 
 module.exports = router;
